Fix memory leak in N1qlQueryConstructor
Memory leak was observed when the JavaScript execution crashed.
Hitherto, when a call to N1qlQuery was made in JavaScript,
N1qlQueryConstructor of C++ would create an empty object and assign
the attributes and methods and return the instance to JavaScript.
If the execution ended abnormally, it caused a memory leak as the
VM shut down and could not reclaim the memory allocated
by the C++ world.
Thus, the fix incorporated in this patch is to defer the instantiation
of N1qlQuery to the JavaScript world. In essence, we are simulating
the "virtual functions" concept of C++ where we define the iter,
execQuery, stopIter methods in the global scope. Since we force
the N1qlQuery to be instantiated in the JavaScript world, if there
is a crash, then all the allocated memory is reclaimed by the VM.

The following discussion on v8-users is relevant -
https://groups.google.com/forum/#!topic/v8-users/XoOj11tpiHE

Change-Id: Ia562ad95e73774a0b23d028613df30504172e6f5
Reviewed-on: http://review.couchbase.org/77026
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: Harsha H S <hhs.couchbase@gmail.com>