[Timers] Prune timer records in xattr, pointing to time in past
For a given document in source bucket, over it's lifetime - it's quite
possible that it has lots of timer events created against it. In
such cases, it makes sense to cleanup of timer event records under
xattr path "eventing.timer" and ascertain that it doesn't contain
events pointing to time in past.

Change-Id: I09b2170289dbb6e548589ab03b405840bdef1f61
Reviewed-on: http://review.couchbase.org/78176
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>