[Debugger] Avoid V8 Debugger to halt processing of Eventing.Consumer
Till now, running instance of V8 Debugger is tied up with an
Eventing.Consumer instance which is responsible for processing a set of
vbuckets. Because of Debugger running, execution of the events landing
on vbuckets owned by Eventing.Consumer is halted. This isn't a desirable
behaviour if we intend debugger to be used in production scenarios.

With this change, a brand new V8 worker instance is spawned just for the
debugging session. This worker instance will get an event to it, which
it will use for debugging session and while debugging is on the other V8
worker instance tied up Eventing.Consumer will process events at it's
full speed. When debugging session is done with first event, then next
sampled event will be passed to it and it will go on with usual
debugging.

Change-Id: I205fac9da49a86bd65581c54c0c83eb896cb3d9d
Reviewed-on: http://review.couchbase.org/82393
Reviewed-by: Gautham B A <gautham.bangalore@gmail.com>
Tested-by: Gautham B A <gautham.bangalore@gmail.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>