[Rebalance] Subset of plasma records didn't transfer during rebalance
This primarily showed up in case host name update. In such case, vbucket
takeover and give up routines ran at same time. Vbucket ownership lead
to cleaning up of dcp connection associated with old hostname of KV node
and as a result of all STREAMEND vbucket processing metadata was updated
as well. This affected the working of vbucket giveup routine which
relied on that in-memory metadata.

This change basically creates a copy of that in-memory structure, so
that vbucket giveup routine is unaffected by behaviour of vbucket
takeover routine.

Change-Id: I1e9f12caea52285b18922b92541fc5a8b4e3c4d4
Reviewed-on: http://review.couchbase.org/85121
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>