[MB-26785] Negative progress for eventing rebalance
Leader eventing node chosen by ns_server is supposed to write the no. of
vbucket ownerships that need to shuffled between eventing node as per
the generated plan. For this it talks all existing eventing nodes that
are participating as part of rebalance.

In this case, some of eventing nodes didn't respond to request from
leader node at the rebalance kick off but they do respond to requests to
later on. As a result the initial count of vbuckets to shuffle around
didn't capture state of the world for the nodes that didn't respond at
rebalance kickoff. This lead to showing -ve progress once the other nodes
start responding.

Change-Id: Ibcfda85a50fde487e731e5667477bf57d3c452e9
Reviewed-on: http://review.couchbase.org/85435
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>