[MB-100] Replace line feed with std::endl
Immediately flushing buffers helps with debugging. So far we've tried to
stick with line feed as it showed some perf benefits when we tested long
back. But now that we've log levels and if user is running debug/trace
level logging, they would probably be okay in paying that perf penalty.

Change-Id: I3d55521f10603ae659706b3a2fc33c9305dde9a1
Reviewed-on: http://review.couchbase.org/86192
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>