[Rebalance] Remote node request for timer data
Source eventing that used to give up ownership of vbuckets, pre-created
the temporary plasma store corresponding to every vbucket it's giving up
ownership for. And then remote node used to grab those timer data in the
temp plasma store corresponding to the vbucket, it's trying to own. This
model worked except for the case of host name update, case when cluster
moves from 1 -> 2 nodes.

In order to mitigate this problem, remote nodes now make RPC calls to
node that was the previous owner of the vbucket to create plasma store
for the corresponding vbucket. This model should work in all cases i.e.
with and without host name update.

Aside from this change, gradually moving the source code towards agreed
upon logging format.

Change-Id: I0cf7b1db87ff7df2d645d35a29ad35da8aa4bff3
Reviewed-on: http://review.couchbase.org/86282
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>