[Rebalance] Update in-memory vbucket states only after metadata writes
State of vbucket is maintained in different places. One is in-memory
cached copy and another is checkpoints maintained in metadata bucket.
Till now, we used to update in-memory cached copy of vbucket states
before metadata writes to couchbase bucket was successful. At times, the
metadata update might not go through or is delayed(probably because
we're retrying bucket op because of some failure). In those cases
in-memory cached copy of vbucket states, especially during rebalance
would get inconsistent with what's in metadata bucket. So to avoid this
problem, this patch tries to update in-memory vbucket info related cache
only after metadata bucket writes go through successfully.

Change-Id: Ic97f897402305071b298e498783a8f5b629b6a3a
Reviewed-on: http://review.couchbase.org/86616
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>