[Rebalance] Avoid race for checkpoint update
Checkpoint metadata so far was updated in 2 different places marking the
vbucket ownership give up. One when DcpCloseStream is issued and another
when DCP_STREAMEND is received for the same vbucket. Reason we added
former was because under load we noticed some DCP_STREAMENDs were missing.

With above model, there is a race. Sample scenario:

n0 - Issues DcpCloseStream for vb 0 and updates metadata
n1 - Sees vb 0 doesn't have an owner and n1 is supposed to own it as per
     the plan, so it takes up the ownership.
n0 - DCP_STREAMEND is received for vb 0, so it updates the metadata
     again, without noticing that n1 has already taken up vb 0
     onwership. This isn't desirable.

Because of above scenario, this changes avoids metadata update on
DcpCloseStream.

Aside from addressing above race, reducing some unnecessary logs. These
additional log messages are noise when we are trying to debug from log
files.

Change-Id: I5a03a1ecf4d9895e716b7d3160f75a3d0bfbbf57
Reviewed-on: http://review.couchbase.org/87179
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>