Store keepNodes from PrepareTopologyChange call to metakv
This assists in cases where eventing-producer gets re-spawned and then
kicks off it's own planner. For correct plan, node requires to know the
current set of valid nodes that are to be part of the cluster. So in
such cases, planner will read from metakv location for list of active
eventing node uuids.

Change-Id: I3b17b5f4e9df8dd979b5348f44f3f113def90566
Reviewed-on: http://review.couchbase.org/87181
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>