[MB-26859] Don't exit metakv callback functions on Eventing nodes...
...that are going to be rebalanced out

As part of change to fail rebalance when one or more apps are undergoing
bootstrap, change was made to exit metakv callback functions for
settings/topology/app code changes if node isn't part of keepNodes
supplied from ns_server. This logic would have problem in-case of
rebalance out of Eventing nodes, because those nodes won't be part of
keepNodes instead would be in list of ejectNodes. The callbacks needs to
be triggered on nodes that are in ejectNodes as they need to give up
vbucket ownerships. Hence this patch adds logic to trigger metakv
callbacks on nodes that are about to be ejected from the cluster.

Change-Id: Ia804341105ec337f24baed00e0942ed89c3f6208
Reviewed-on: http://review.couchbase.org/91058
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>