MB-27813 Restream vbuckets even if Eventing rebalance is yet to be triggered
Eventing-producer could receive DCP_STREAMEND because of 2 cases:

* Vbucket movement across data nodes in the cluster as part of KV
  rebalance.
* Vbucket ownership related changes across Eventing nodes in the cluster
  & even between Eventing.Consumer instances running on same Eventing
  node.

Till now, STREAMEND that eventing-producer used to receive because of KV
rebalance - eventing-producer didn't try to reclaim the ownership back
by issuing STREAMREQ against it. Instead, it waited till ns_server
notified it that it should kick-off Eventing rebalance. And it would
issue STREAMREQ for all pending vbuckets that needs to be re-streamed.

This approach didn't work well on performance cluster where DCP message
throttling also came into picture. In general, this problem would
surface with slow handlers where back-off from eventing-consumer has been
kicked off.

Instead, now in this change - eventing-producer issues STREAMREQ for
vbuckets that it receives STREAMEND & as per plan if the
Eventing.Consumer is supposed to own it. This would address the problem
of rebalance hang that surfaced with back-off in-place.

Change-Id: Ief44c886eb4e505cf5026ebccba3cdd585bb9b31
Reviewed-on: http://review.couchbase.org/91347
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>