MB-28968 Separate out handling of messages on failover log channel
In current code, we expected number of messages on failover log channel
to be equal to number of STREAMREQ that are issued by Eventing.Consumer.
But that may not be the case now, as we now also pipeline feedback from
STREAMEND received in-cases when DcpRequestStream is issued and then KV
node goes down - there instead of receiving response from DCP producer
w.r.t. DCP_STREAMREQ call, dcp-client sends a fake DCP_STREAMEND.

Change-Id: I08533fad78d6215df8fbda7939a10dbe7d57bdac
Reviewed-on: http://review.couchbase.org/92158
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>