[CI-Rebalance] Use separate maps to represent deployed apps
There are 2 different paths, that are interested in knowing the set of
currently deployed apps:

(a) End user & UI, who are querying via REST endpoint /getDeployedApps
(b) Inside server, Eventing.Consumer needs knowledge about it as well.
    With that knowledge it acts upon STREAMEND that it sees from KV. For
    deployed app it will fire STREAMREQ & for undeployed apps it will
    discard the STREAMEND and do nothing.

For (a) - it should be notified only when all the metadata in bucket and
on local plasma store is purged.

For (b) - it's time sensitive i.e. it can't wait till all cleanup
activity that (a) expects is wrapped up.

Hence exposing 2 different maps that will be leveraged by (a) & (b)
respectively.

Change-Id: I888f07f6ca5badb02f4d25dac3801bfd87a77c18
Reviewed-on: http://review.couchbase.org/92614
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>