MB-29260 Mitigate race for vbucket ownership
During KV+Eventing rebalance, in some cases it's possible vbuckets to
restream(as a result of KV rebalance) wouldn't get accounted for
vbuckets to give up as per the generated plan.

To mitigate this, post every successful STREAMREQ response - we would
check if current eventing worker is supposed to own it or not. If not,
then it would issue a close stream and update metadata about giving up
ownership.

Change-Id: Ie2c2904c9816c7a1f7df3862877795a45b30c49f
Reviewed-on: http://review.couchbase.org/92839
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>