MB-29349 Add stats counter to track vbucket seq nos
Motivation:
----------
During volume tests, we noticed some events getting missed. And
with current set of stats it's hard to debug which vbucket lot
the events during rebalance - as stats are cumulative presently.

Changes:
--------
Adding stats for tracking dcp related sequence nos to track as and
when vbuckets are migrated to different eventing nodes as part of
rebalance.

Change-Id: Ibe8ae94eec714da634713add52f1ce9df6b65fcc
Reviewed-on: http://review.couchbase.org/93345
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>