MB-100: Do not allow excess worker counts
In the UI, we say one can have 1024 workers, and validate that
it is less than 128. But in testing, things are rarely advisable
to go beyond handful of workers. Reduce the number to 32, and also
validate it at that. This is a UI only change, if there is some
use case that can use more counts, it can be changed via REST APIs.

Change-Id: I4edae366cdaa0de820ad2bbaad9654cd6f61706b
Reviewed-on: http://review.couchbase.org/94666
Reviewed-by: CI Bot
Reviewed-by: Sriram Melkote <siri@couchbase.com>
Tested-by: Sriram Melkote <siri@couchbase.com>