MB-27158: Populate the error back to eventing.log when underlying service throws error
When there is no query node in the cluster and the event handler code
contains a N1QL query, no exception is thrown and the target buckets
are getting updated with empty results. This is fixed with this commit
and an exception is thrown.

Also, the exception is logged as an object and not as string. To make
it more readable, the exception object is stringified and then dumped
into the log

Change-Id: I8b79f8fc45c09ab87e42b0d576815c99eb2caca3
Reviewed-on: http://review.couchbase.org/98675
Reviewed-by: CI Bot
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>