MB-30764 Force syncSpan() for timer stores after rebalance
During rebalance, if the timer span is updated on one node (say n1) and the timer event is
supposed to get executed on another node (say n2), then the node n2 may not read the updated
span as the dirty bit in n2's store is not set to true.

This results in missing of some timer events as node n2's timer store contains the older span.
To avoid this scenario, we force all the timer stores to sync span when rebalance is in progress

This patch fixes the issue with large context size in timers

Change-Id: I887b3769a17924e09f1208eaf8ee77997e6fdf7b
Reviewed-on: http://review.couchbase.org/99755
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Reviewed-by: Sriram Melkote <siri@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>