MB-28052[13/n] Contain duplicate events during rebalance
Motivation:
==========

With the introduction of support for from_prior deployment, there was
jump in duplicate events during rebalance - specifically if user had
deployed with "everything" as dcp feed boundary.

Changes:
=======

To address the regression in duplicate events during rebalance, changes
done:

* Tracking STREAMREQ for a vb is done is issued for the very first time
  post bootstrap.
* Once STREAMREQ success response comes back from DCP producer, we
  update in checkpoint blob saying "initial stream req as part of
  bootstrap is finished".
* Above flag would allow subsequent would now allow any new STREAMREQs
  to choose STREAM START SEQ as last processed seq_no from last eventing
  node.

With previous from_prior deployment changes, there was no away to
differentiate a checkpoint blob for vbucket with no owner is because of:

(a) STREAMEND received as part of eventing rebalance, resulting in it's
    previous owner eventing node stamping that it gave up it's
    ownership.
(b) STREAMEND received as part of pausing the execution of Function.

This patch would allow us to differentiate b/w (a) & (b), and hence
contains duplicate event during rebalance.

Change-Id: I915849c8793aca089aa5218a12c03a319db91686
Reviewed-on: http://review.couchbase.org/101173
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>