MB-35047: Adding nil check to mitigate golang panic
If the old handlers don't contain any data
in the credentials path during upgrade, they
can cause golang panic during parsing flatbuffer
payload

The other problem addressed is the deleting of
curl bindings, that caused a similar golang
panic due to the curl credentials length being
lesser than the ones stored in the apps path.
We trim the curl bindings to the length of the
curl credentials in the credentials path.

Change-Id: I8f2146102659753b875421c40a829862743eb95f
Reviewed-on: http://review.couchbase.org/115552
Reviewed-by: Gautham B A <gautham.bangalore@gmail.com>
Tested-by: Suraj Naik <suraj.naik@couchbase.com>