MB-36822: Settings should be changeable for deployed apps
Settings change for a deployed app post to the settings
endpoint with the default stream_boundary which is
"from_everything", but due to the recent additional
check in the set Settings which does not allow change
of stream boundary for deployed apps causes the settings
change to fail. This change fixes it by removing the
stream boundary parameter before posting to the endpoint.

Change-Id: Ie78078c3e7f71dac9a03ceb6bd22bc22030cbc71
Reviewed-on: http://review.couchbase.org/117629
Reviewed-by: Gautham B A <gautham.bangalore@gmail.com>
Tested-by: Suraj Naik <suraj.naik@couchbase.com>