Disable 'Delete' button MB-37854, Improve Editor button labels MB-37867
Disable 'Delete' button when Handler state is != 'undeployed'
or going through a change MB-37854

Toggle 'Edit JavaScript' button's label to 'View Javascript' when
state is not in ( 'undeployed' or 'paused' ) MB-37854

This will prevent a "Delete" action from being selected (and acted
upon) during an undeploy prior to the undeployed state actually
being reached (or a busy delayed response on a 30 bucket system)
Without this fix a refesh is sometimes needed to discover the
Function was not actually deleted.

Toggle the label 'Edit JavaScript' between 'View Javascript'
makes it very clear when edits can be saved in the Function's
JavaSript code editor.

Change-Id: Ia39d746268239a8a30e938ee0c687630d354cf1d
Reviewed-on: http://review.couchbase.org/121547
Reviewed-by: Gautham B A <gautham.bangalore@gmail.com>
Tested-by: Jon Strabala <jon.strabala@couchbase.com>