MB-32437:Add ServiceChangeNotifier to cache Cluster Info
This patch adds ServiceChangeNotifier to notify on a channel
in case of any service changes.

The cluster info cache and the bucket refresh is done only
in case of a notification from the service change notifier
or periodically. The periodic refresh timer is configurable.

This speeds up the call to KVVbmap and the bucket refresh is
replaced with calls to GetBucket which returns the refreshed
bucket from the Supervisor's bucket map.

Made the cache in util/cluster_info.go a sigleton per
process and utilized FetchClusterInfoClientCache in 11
key places across 3 files

(in: service_manager/http_handlers.go)
  determineCurlWarning
  isMixedModeCluster
  checkVersionCompat

(in: service_manager/validation.go)
  validateBucketExists
  validateNonMemcached

(in: util/util.go)
  KVVbMap
  NsServerNodesAddresses
  KVNodesAddresses
  EventingNodesAddresses
  CurrentEventingNodeAddress
  CountActiveKVNodes

Replaced the calls to bucket Refresh with GetBucket which
returns the buckets from the supervisor's bucket map

Change-Id: I13f6f9262564237b26753f9d9625a8a530f2d356
Reviewed-on: http://review.couchbase.org/121533
Reviewed-by: Suraj Naik <suraj.naik@couchbase.com>
Tested-by: Suraj Naik <suraj.naik@couchbase.com>