Protect against importing "dcp_stream_boundary": "from_prior" MB-38020
There can be exported functions in the wild that have

    "dcp_stream_boundary": "from_prior",

These functions wil import but silently fail staying in
the undefined "..." state because they are no longer legal
as the POST to save the import into the system fails when
leaving the 'settings' dialog and and entering the code
editor. You will see the HTTP request to import fail in
the JavaScript console.

    POST http://<ip>:8091/_p/event/api/v1/import 400 (Bad Request)

Eventually the UI logic (after 10 status reads) will remove
the Eventing function that errored and didn't really succesfully import.

The fix is pretty basic just like we set (on every import)

    creationScope.appModel.settings.processing_status = false;
    creationScope.appModel.settings.deployment_status = false;

we also need to also set

    "dcp_stream_boundary": "everything",

The fix is pretty basic just like we set (on every import)

    creationScope.appModel.settings.processing_status = false;
    creationScope.appModel.settings.deployment_status = false;

we also need to also set

    "dcp_stream_boundary": "everything",

Change-Id: I9b095b19b23d441551a44768dc66460e9802fe3d
Reviewed-on: http://review.couchbase.org/122674
Reviewed-by: Gautham B A <gautham.bangalore@gmail.com>
Tested-by: Jon Strabala <jon.strabala@couchbase.com>