MB-25475: prepare Eventing service to lazy loading
In order to prepare it to lazy loading.

This change introduces new main.js module which
is a root module of Eventing UI and place for
Eventing UI inital config.

In this change, we also indicate that the main.js
will be imported into generated pluggable-uis.js
(more here http://review.couchbase.org/c/ns_server/+/125551)

Change-Id: I790835d177fb1d1cec84f71edc5c2e8d03c4a33d
Reviewed-on: http://review.couchbase.org/c/eventing/+/125783
Tested-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>