MB-38787: Assign vbuckets to eventing node for given bucket
Eventing tries to get the kv node vbucket distribution for a bucket.
If bucket is not present/deleted it will throw error and return. When
source bucket gets deleted it will call this function to distribute
vbuckets to cleanup. For cleanup, it should be metadata bucket. Since
source bucket is not present it will return error and skip cleanup of
metadata bucket.

Change-Id: I94ebdec015bfcfbf2679afd22bb11c23962de212
Reviewed-on: http://review.couchbase.org/c/eventing/+/125987
Reviewed-by: CI Bot
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Tested-by: <ankit.prabhu@couchbase.com>