MB-38297: cap queue sizes at 61MB to smoothen throttling as new
handlers get deployed. For e.g, when eventing quota is 3GB and only
one function is deployed queues size can bloat upto 500MB. Now when a
second function is deployed, queue cap suddenly drops to 250MB. This
will cause prolonged period of throttling until queue size comes down
to 250MB or less. In case of slow functions this can take hours.
And we will not be reading mutations from DCP conenctions in this
case causing KV to back up. Which in turn causes any further new
handlers that get deployed to not recieve mutations for hours.

Change-Id: I4491fd652bd239da1b44599a52da99d34f1a2004
Reviewed-on: http://review.couchbase.org/c/eventing/+/128603
Tested-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Reviewed-by: CI Bot
Reviewed-by: <ankit.prabhu@couchbase.com>