MB-42554: Prevent Delete button from staying blocked when the source namespace is deleted
All UI Elements (should) rely on app.status to display sync/async lifecycle information and/or allowing/disabling button.
Meanwhile all the settings are treated as write only and are not updated when there is a change in the store. Migrating all
actionable elements in the UI to app.status will keep the behaviour consistent. The settings are only fetched when there is
a page refresh.

Change-Id: I1437fa872703047405f43440c8e0a1a5253013b8
Reviewed-on: http://review.couchbase.org/c/eventing/+/142735
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Tested-by: Vinayaka Kamath <vinayaka.kamath@couchbase.com>