MB-45967 : Retry on DML Failures
It has been seen that a failure on DML N1QL query (such as TMPFAIL,
ENOMEM, NMVB, Connection error, SOCKET_SHUTDOWN etc) are encapsulated as
DML_FAILURE by lcb which is not a retriable / transient error. This can
then result in document mismatch for N1QL handlers. Hence, we need to
retry on DML_FAILUREs too.

Change-Id: I3aef901f4cf283c759394ea8b1b35fe6708f7aa2
Reviewed-on: http://review.couchbase.org/c/eventing/+/152666
Tested-by: <abhishek.jindal@couchbase.com>
Reviewed-by: <ankit.prabhu@couchbase.com>
Reviewed-by: CI Bot