MB-45837 : Open failover log DCP stream for only the required vbucket
Currently the consumer is opening up DCP streams on behalf of all 1024
vbuckets for failover logs everytime when the requirement is to open up
stream only for the vbucket passed in argument. This causes a lot of
delays in terms of opening up and shutting down 1024 connections
everytime which is has also proved to have caused increased rebalance
times.

Change-Id: I4e000d6b3a2dbcc9fb562460bbf801651779e752
Reviewed-on: http://review.couchbase.org/c/eventing/+/152685
Tested-by: <abhishek.jindal@couchbase.com>
Reviewed-by: <ankit.prabhu@couchbase.com>
Reviewed-by: CI Bot