MB-46543 : Append vb to vbsRemainingToRestream without checking for its presence
Problem:

If a DCP stream request for a vbucket fails down the line at the time
Planner is still running, there can be a case where this partition is
not pushed to the vbsRemainingToRestream list. Hence, the controlRoutine
won't retry DCP Stream Request resulting in rebalance to be stuck for
the next 20 minutes.

Solution:

No need to check for presence of vb in vbsRemainingToRestream list.
Appending the vbsRemainingToRestream list with a vbucket *that is
already present in the list* is actually indicative of the fact that the
initial stream request failed which is why there is a need for
requesting stream for this partition again.

Change-Id: Ib06d9b13e7632105e926eca6354e8126e7eb3410
Reviewed-on: http://review.couchbase.org/c/eventing/+/154427
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Reviewed-by: CI Bot
Tested-by: <abhishek.jindal@couchbase.com>