MB-45967 : Disambiguate between retriable and non-retriable DML errors
As per the discussion in the MB, for the time being we are adding a
filter to further disambiguate between a retriable DML error and
non-retriable.
That said, once this logic is added to libcouchbase in future CB
versions, this check won't be needed here in eventing.

Please refer:
https://github.com/couchbase/go-couchbase/blob/1d32284da76dab362bda3b1775372bebc755412e/client.go#L194
This section of go-couchbase (called by query) deals with retry
mechanism for all transient errors such as  nmvb, tmpoom, connection
error etc.
It is guaranteed that ONLY such transient errors will be encapsulated
under a string "unable to complete action after X retries".
This is why for the time being eventing can leverage this property to
disambiguate between retriable and non-retriable.

Change-Id: I2b81a990858aeab1380eb7faba4edef4cef18b3c
Reviewed-on: http://review.couchbase.org/c/eventing/+/154646
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Reviewed-by: CI Bot
Tested-by: <abhishek.jindal@couchbase.com>