MB-48117: Fix missing event due to handling of close dcp request
Currently handlePacket will also handles the control message
and response from the dcp producer. If control message is CloseStream
it won't send the event to eventing consumer. If the packet contains
error response for streamReq due to this race condition it will skip
sending this response to eventing consumer and consumer will wait for
the response.
After this fix genServer will manage handling of response from
dcp producer and control messages.


Change-Id: I248f0990ba5ee1887c962496ea737b6ee22c4a8d
Reviewed-on: http://review.couchbase.org/c/eventing/+/160150
Reviewed-by: <abhishek.jindal@couchbase.com>
Reviewed-by: <varun.velamuri@couchbase.com>
Reviewed-by: prataprc <prataprc@couchbase.com>
Reviewed-by: CI Bot
Tested-by: <ankit.prabhu@couchbase.com>