MB-49418: Check cached entry for app state
When delete and deploy is executed at same time and there is a
delay on metakv callback, eventing will delete the handler and deployment
will get stuck.
With this change Delete will check for any deploy request for app
executed on the same node. If app is in undeployed state then only
delete the app

Change-Id: I6221eb5ee3f5563e549b8e0eca64d9a62c6b1453
Reviewed-on: https://review.couchbase.org/c/eventing/+/166409
Reviewed-by: CI Bot
Reviewed-by: <abhishek.jindal@couchbase.com>
Tested-by: <ankit.prabhu@couchbase.com>