MB-50216: Refresh cluster info cache when validation fails
If node gets added after pools default call and before pool service call
then thisNode will be in different index position. This position is used
to determine the address of this node service and to determine which all
vbs it owns. Changing the position can lead to owning vbs of other
eventing node

Change-Id: I212ec418ae34d56fd1ac303dabb8b2060b86aa70
Reviewed-on: https://review.couchbase.org/c/eventing/+/168134
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Reviewed-by: CI Bot
Tested-by: <ankit.prabhu@couchbase.com>