MB-50456 : Proactively stream encryption level changes to C++ consumer
In an attempt to make decisions in C++ consumer based on
the changes in encryption level, (maybe handling timer store failure)
or (for future), auto-repairing lcb handles we need a mechanism to stream
changes in encryption levels to the consumer process.

As of this patch, we are making a decision to either actively bail out
or not out of the failing timer store operations on encryption level being changed
from plain text to tls only.

Change-Id: I00269b7384a100d8c945927835557785a557bd14
Reviewed-on: https://review.couchbase.org/c/eventing/+/168711
Tested-by: <abhishek.jindal@couchbase.com>
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Reviewed-by: CI Bot