MB-50284 : Proactively stream encryption level changes to C++ consumer
In an attempt to make decisions in C++ consumer based on
the changes in encryption level, (maybe handling timer store failure)
or (for future), auto-repairing lcb handles we need a mechanism to stream
changes in encryption levels to the consumer process.

As of this patch, we are making a decision to either actively bail out
or not out of the failing timer store operations on encryption level being changed
from plain text to tls only.

Change-Id: Ie97cd0057b79d9facce61f89221771acd118c73a
Reviewed-on: https://review.couchbase.org/c/eventing/+/168751
Tested-by: <abhishek.jindal@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: <abhishek.jindal@couchbase.com>