MB-50780: Check keyspace exist when handler is not in running apps
Eventing puts the bucket in watcher list and continue to spawn the app.
Supervisor puts the app into running Apps after app is completely spawned.
Current code undeploys the app whenever there is a change in manifest uid
and app is not in running apps without checking deletion of scope/collection.
Solution is to check for keyspace deletion before undeploy the app

Change-Id: I55b9329dbf3f51f6d5f62a5d7cd23bdcebd5823c
Reviewed-on: https://review.couchbase.org/c/eventing/+/170152
Reviewed-by: CI Bot
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Tested-by: <ankit.prabhu@couchbase.com>