MB-51390 : Proactively stream encryption level changes to C++ consumer
In an attempt to make decisions in C++ consumer based on
the changes in encryption level, (maybe handling timer store failure)
or (for future), auto-repairing lcb handles we need a mechanism to stream
changes in encryption levels to the consumer process.

As of this patch, we are making a decision to either actively bail out
or not out of the failing timer store operations on encryption level being changed
from plain text to tls only.

Change-Id: I866094a22a607de586b5f0221e3b188a8e9c7094
Reviewed-on: https://review.couchbase.org/c/eventing/+/173249
Tested-by: <ankit.prabhu@couchbase.com>
Reviewed-by: <ankit.prabhu@couchbase.com>
Reviewed-by: CI Bot
Well-Formed: Restriction Checker