MB-51442: Fix issue to enable user to modify function settings if s/he has Manage Scope Functions role.
The save button for a non Full Admin user was disabled this prevented
changing function settings even though the use had Manage Scope Functions
role. This was because the Listen to Source fields and Metadata fields
were not filled as the user lacked permissions to the source and meta-
data Keyspaces.

This fix pre-fills deployment Keyspace details (source and metadata
Keyspace details) by default to enable Save button so that user can
change Settings parameters on the Settings screen.

Change-Id: Ia942de5c1ea22cd55e6b8487a5a50d32df1e15f1
Reviewed-on: https://review.couchbase.org/c/eventing/+/176147
Tested-by: Mohammad Nawazish Khan <nawazish.khan@couchbase.com>
Reviewed-by: abhishekjindal <abhishek.jindal@couchbase.com>