MB-52588: Add manifest id while recreating checkpoint blob
dcp stream request expects manifest id to be sent with the request.
Otherwise it will error out with EINVAL.
This patch will include manifest id in the checkpoint blob from the
in memory stats and otherwise populate it with "0" which is the
beginning of the manifest id. So while making request eventing will
always send non empty manifest id.

Change-Id: Iead6f171f27183c214d2cf8d6130a8c174ccf11a
Reviewed-on: https://review.couchbase.org/c/eventing/+/176361
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Reviewed-by: CI Bot
Tested-by: <ankit.prabhu@couchbase.com>