MB-100: make `itemCount` constant
I've changed the identifier `itemCount` from a variable to a constant,
as we use it to avoid magic literals (i.e. improve code readability).

Additionally, using constants makes it clear that the value of the
identifier will not change; it also allows the compiler to perform
optimizations such as inlining.

References:
1. https://blog.boot.dev/clean-code/constants-in-go-vs-javascript-and-when-to-use-them/
2. https://blog.learngoprogramming.com/learn-golang-typed-untyped-constants-70b4df443b61
3. https://www.sobyte.net/post/2021-06/const-in-golang/

Change-Id: Ie2d5fa290421b1d29c126ec57ceae77e29e8387c
Reviewed-on: https://review.couchbase.org/c/eventing/+/188724
Reviewed-by: CI Bot
Reviewed-by: <ankit.prabhu@couchbase.com>
Tested-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>