MB-59796: Ignore documents present in `_system` scope
Eventing functions listening to wildcard collections, such as
`<bucket_name>.*.*`, should ignore documents present in `_system` scope.
Any mutation of documents in the `_system` scope should not trigger the
eventing function.

Additionally, we disallow listening to any collection with `_system`
scope as this scope is only meant for internal use and not for end
users.
The "Function Scope", "Listen To Location" and "Eventing Storage" values
of an Eventing function cannot use the `_system` scope. This change has
been implemented in the Web UI and the REST APIs.

Change-Id: I4f233ff48c3adbff2f6677cee200d225e83e11df
Reviewed-on: https://review.couchbase.org/c/eventing/+/202091
Tested-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>
Reviewed-by: CI Bot
Reviewed-by: <ankit.prabhu@couchbase.com>