MB-61243: Eventing requires unnecessary rebalance
The `service manager` locally stores a list of active eventing
nodes of the cluster in the Go layer, which is used by the
`stateToTopology` function for providing information about
the current topology of eventing in the cluster. When we restart
a running cluster or kill the eventing-producer process,
the `service manager` also gets respawned, and this list is
not updated correctly with the eventing nodes present in the cluster.
This results in ns_server thinking that the eventing's node
topology has changed, thus enforcing an unnecessary rebalance
of an already balanced cluster.

This fix retrieves the current list of active eventing node
UUIDs from MetaKV, and correctly updates it in the constructor
of the `service manager` for avoiding any undesired rebalance.

Change-Id: I44c733934fc6b16c30056e18824c069c0018e752
Reviewed-on: https://review.couchbase.org/c/eventing/+/209009
Reviewed-by: <ankit.prabhu@couchbase.com>
Tested-by: Hiren Bavaskar <hiren.bavaskar@couchbase.com>
(cherry picked from commit a5e9dc7e88c923c8f0b39642270380c343824e68)
Reviewed-on: https://review.couchbase.org/c/eventing/+/209206
Well-Formed: Restriction Checker
Tested-by: <ankit.prabhu@couchbase.com>