MB-62863: Align enums used for communication between C++ and Go
A key part of the communication between the Go and C++ processes happens
using `event` and `opcode` values located in the header of the exchanged
messages.
Currently, the code that defines and interacts with these values
contains unnecessary code which we can remove.

Additionally, the code is simplified by adopting `enum` and `enum class`
in Go and C++, respectively.

Finally, in the C++ layer, there were a lot of if-else-if ladders to
convert and decode the messages from the Go layer. These ladders have
been removed and replaced by `assert` statements that we can enable as
required to help catch bugs while working on new features.
Note: These `assert` statements are disabled in production code, so they
do not affect the service's performance.

Overall, this patch will improve the code's readability,
maintainability, safety and resource efficiency.

Change-Id: Ifa0a1d313f07c073ec5abb31f533545e73c13220
Reviewed-on: https://review.couchbase.org/c/eventing/+/213364
Reviewed-by: <ankit.prabhu@couchbase.com>
Tested-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>