MB-29219: Implement OnDeploy in Go layer
To run OnDeploy on a single v8 worker, we need to choose a
function handler of a given node to send OnDeploy event to
the C++ layer. In this patch, we design this election in
the following two steps:

* The first function handler of each node is marked as leader
* Choose one node via checkpoint insertion in the metadata
collection of the function

The (function handler, node) pair forms the OnDeploy leader,
which creates the runtime environment essential for the
utility worker that executes OnDeploy.

To ensure that OnDeploy completes before starting DCP streams,
we poll all the function handlers on all nodes until we
receive the final status (Finished/Failed) from C++ layer
via the feedback communicator. Based on the status:

* `Finished`: Proceed to deployment of the function
* `Failed`: Revert back to previous state of function
(Paused/ Undeployed)

We persist the OnDeploy checkpoint in the metadata collection
and delete it only before next deployment, to avoid race
conditions during OnDeploy failure cases.

Change-Id: Id8eadbc2f2fe616facdd2317b1350cca6f170037
Reviewed-on: https://review.couchbase.org/c/eventing/+/219132
Reviewed-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>
Reviewed-by: <ankit.prabhu@couchbase.com>
Tested-by: Hiren <hiren.bavaskar@couchbase.com>